requirement.txt
Django>=3.2,<4.1
djangorestframework>=3.11,<4.0
djangorestframework-authtoken>=1.1,<2.0
psycopg2-binary>=2.8,<3.0
django-allauth>=0.41.0,<1.0
django-rest-auth>=0.9.5,<1.0
